sync: add back RwLockWriteGuard::map and RwLockWriteGuard::try_map (#3348)
